style: work on base structure and more
add date-start and date-end

change 'Things that I got help from to make the project' to 'Reference - Things that I got help from to make the project'

This page doesn't have a good style because I don't have enough experience and I was in a hurry.

In the first part of references, I was use image and link for better viewing.