fix: Remove duplicate files
By copying the projects, I also copy all the references, which include all the pictures, etc.

I copied the practice project of sessions 5 and 6 from session 4, and in the meantime, all the reference files that have a significant volume are also copied.

So, I deleted the reference files in sessions 5 and 6, and in their place, I added a link to reference session 4 with the necessary explanations.